Refactor ExchangeAIAnalysis component by removing unused imports and components, and simulate AI insights for market analysis. Updated styles in ExchangeAIAnalysis and StockSearchSection to remove unnecessary imports.
