+++ src/containers/analysis-page/exchange-ai-analysis-section/ExchangeAIAnalysis.module.scss
@@ -1,5 +1,3 @@
-@import '../../../styles/breakpoints.scss';
-
 .container {
   width: 100%;
   padding: 1.5rem;
