+++ src/containers/analysis-page/stock-search-section/StockSearchSection.module.scss
@@ -1,5 +1,3 @@
-@import '../../../styles/breakpoints.scss';
-
 .container {
   width: 100%;
   max-width: 800px;
